pyproject.toml
[project]
name = "chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "datasets>=4.0.0",
    "fastapi>=0.116.1",
    "itsdangerous>=2.2.0",
    "jinja2>=3.1.6",
    "numpy<2.0.0",
    "openai>=1.99.3",
    "pyjwt>=2.10.1",
    "pymysql>=1.1.1",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "scikit-learn>=1.7.1",
    "sqlalchemy>=2.0.43",
    "transformers>=4.55.0",
    "uvicorn>=0.35.0",
]
